+++ database/migrations/2018_09_10_084820_create_user_for_migrations.php
@@ -13,36 +13,38 @@ class CreateUserForMigrations extends Migration
      */
     public function up()
     {
-        $errors = [];
+        if (!env('IS_TOOL')) {
+            $errors = [];
 
-        $password = env('SYSTEM_PASSWORD');
+            $password = env('SYSTEM_PASSWORD');
 
-        if (!empty($password)) {
-            try {
-                DB::table('users')->insert([
-                    'username'      => 'migrate_data',
-                    'password'      => bcrypt($password),
-                    'email'         => '',
-                    'firstname'     => '',
-                    'lastname'      => '',
-                    'add_info'      => '',
-                    'is_admin'      => 1,
-                    'active'        => 1,
-                    'approved'      => 1,
-                    'api_key'       => 'f93d3beb-681f-427f-b608-5ea4d2635cd6',
-                    'hash_id'       => '',
-                    'created_at'    => gmdate('Y-m-d H:i:s'),
-                    'created_by'    => 1,
-                ]);
-            } catch (\Illuminate\Database\QueryException $ex) {
-                $errors[] = $ex->getMessage();
+            if (!empty($password)) {
+                try {
+                    DB::table('users')->insert([
+                        'username'      => 'migrate_data',
+                        'password'      => bcrypt($password),
+                        'email'         => '',
+                        'firstname'     => '',
+                        'lastname'      => '',
+                        'add_info'      => '',
+                        'is_admin'      => 1,
+                        'active'        => 1,
+                        'approved'      => 1,
+                        'api_key'       => 'f93d3beb-681f-427f-b608-5ea4d2635cd6',
+                        'hash_id'       => '',
+                        'created_at'    => gmdate('Y-m-d H:i:s'),
+                        'created_by'    => 1,
+                    ]);
+                } catch (\Illuminate\Database\QueryException $ex) {
+                    $errors[] = $ex->getMessage();
+                }
+            } else {
+                $errors[] = 'SYSTEM_PASSWORD not entered in config';
             }
-        } else {
-            $errors[] = 'SYSTEM_PASSWORD not entered in config';
-        }
 
-        if (!empty($errors)) {
-            dd($errors[0]);
+            if (!empty($errors)) {
+                dd($errors[0]);
+            }
         }
     }
 
@@ -53,8 +55,10 @@ class CreateUserForMigrations extends Migration
      */
     public function down()
     {
-        try {
-            DB::table('users')->where('username', 'migrate_data')->delete();
-        } catch (\Illuminate\Database\QueryException $ex) {}
+        if (!env('IS_TOOL')) {
+            try {
+                DB::table('users')->where('username', 'migrate_data')->delete();
+            } catch (\Illuminate\Database\QueryException $ex) {}
+        }
     }
 }
